requirements.txt
unigradicon 
numpy
pynrrd
h5py
SimpleITK
matplotlib
scipy
pandas